Written lift.cc
-Changed person to unique_ptr<int>
-Lift now stores direction (int, {0, -1, 1})
-Direction flags only in dispatcher
-Added converter (direction to direction_flags)
-Changed event struct to 3 ints (now w\o person, not enough data in him)
-Removed person_states (didn't actually need them)
-Added lift::order
-Added LIFTSPEC structure - for starting values
-Not-so-needed change from size_t to long

+++ ticker.cc
@@ -11,8 +11,9 @@ void timed_obj::wait(int time, wait_type type, bool priority)
 
 void ticker::start()
 {
-    for (size_t time = 0; !waiters_list.empty(); time++)
-    {
+    for (long time = 0; !waiters_list.empty(); time++)
+    {   
+        //Process waiters list
         for (auto it = waiters_list.begin(); it != waiters_list.end();)
         {
             if (it->time < 0) throw std::out_of_range("Waiter's time < 0");
@@ -20,6 +21,7 @@ void ticker::start()
             bool time_out = false;
             switch(it->type)
             {
+                //wait_for waiters
                 case WAIT_FOR:
                 time_out = (it->time == 0);
                 it->time--;
